Extract ResponseUtils class from EmailReceiver
Contains utilities for extracting information from a formatted email
string.
